requirements.txt
python-dotenv
discord.py
pillow
pandas